Action Platform UI/Validation Adjustments (#629)
* Adds Action Platforms Menu to Dashboard

* Changes delete rule for subscriptions: RESTRICT

* Applies policies to the Platform List view

* Migrates platforms.description to string(2000) from text, applies length validations, correct spec for referential integrity

* Applies policy to the controller actions

* Applies max length to the Platform edit/new forms

* Adds improved (?) approve/revoke icons

* Applies the policy to the subscriptions controller

* Applies new signature to policy usage

* Adds action platform manager to the list of visible roles for a ungc user

* Addresses changed exceptations with optimized query

* Updates Lead “tier” title. Closes #632, #625, #624

+++ app/controllers/admin/action_platform/platforms_controller.rb
@@ -1,5 +1,8 @@
 class Admin::ActionPlatform::PlatformsController < AdminController
   before_filter :no_organization_or_local_network_access
+  before_filter :set_policy
+
+  attr_reader :policy
 
   def index
     @platforms = ::ActionPlatform::Platform
@@ -9,11 +12,17 @@ class Admin::ActionPlatform::PlatformsController < AdminController
 
   def show
     @platform = load_platform
+    @liquid_layout = true
+
+    unless policy.can_view?
+      raise ActionController::MethodNotAllowed
+    end
+
     @subscriptions = ::ActionPlatform::Subscription.
-      includes(:platform, :contact, organization: [:organization_type]).
-      joins(:order).
-      where(platform: @platform).
-      order("action_platform_orders.updated_at DESC")
+    includes(:platform, :contact, organization: [:organization_type]).
+    joins(:order).
+    where(platform: @platform).
+    order("action_platform_orders.updated_at DESC")
   end
 
   def new
@@ -21,10 +30,16 @@ class Admin::ActionPlatform::PlatformsController < AdminController
   end
 
   def edit
+    unless policy.can_edit?
+      raise ActionController::MethodNotAllowed
+    end
     @platform = load_platform
   end
 
   def create
+    unless policy.can_create?
+      raise ActionController::MethodNotAllowed
+    end
     @platform = ::ActionPlatform::Platform.new(platform_params)
     if @platform.save
       redirect_to admin_action_platform_platforms_path, notice: 'Platform created.'
@@ -34,6 +49,9 @@ class Admin::ActionPlatform::PlatformsController < AdminController
   end
 
   def update
+    unless policy.can_edit?
+      raise ActionController::MethodNotAllowed
+    end
     @platform = load_platform
     if @platform.update(platform_params)
       redirect_to admin_action_platform_platforms_path, notice: 'Platform updated.'
@@ -44,6 +62,9 @@ class Admin::ActionPlatform::PlatformsController < AdminController
 
   def destroy
     platform = load_platform
+    unless policy.can_destroy?
+      raise ActionController::MethodNotAllowed
+    end
     if platform.destroy
       flash[:notice] = "Platform removed."
     else
@@ -66,4 +87,9 @@ class Admin::ActionPlatform::PlatformsController < AdminController
       :discontinued
     )
   end
+
+  def set_policy
+    @policy = ActionPlatformPolicy.new(current_contact)
+  end
+
 end
